+++ frontend/src/components/admin/pages/AdminProfilePage.tsx
@@ -41,6 +41,7 @@ import ChangePasswordModal from '../modals/ChangePasswordModal';
 import ChangeRoleTitleModal from '../modals/ChangeRoleTitleModal';
 import DeleteAccountModal from '../../shared/DeleteAccountModal';
 import { useMediaCacheControl } from '../../../hooks/useMediaCacheControl';
+import BiometrieSchalter from '../../shared/BiometrieSchalter';
 
 const AdminProfilePage: React.FC = () => {
   const { pageRef, presentingElement } = useModalPage('admin-profile');
@@ -317,6 +318,10 @@ const AdminProfilePage: React.FC = () => {
                     </div>
                   </div>
                 </IonItem>
+
+                {/* Biometrische Anmeldung (blendet sich selbst aus, wenn das
+                    Geraet keine eingerichtete Biometrie hat) */}
+                <BiometrieSchalter variante="users" />
 
                 {/* Medien-Cache leeren */}
                 <IonItem